finished base classes

+++ CrochetApp/backend/Domain/Model/Hook.cs
@@ -9,10 +9,10 @@ namespace CrochetApp.backend.Domain.Model
     public class Hook
     {
         public int Id { get; set; }
-        public float Size{ get; set; }
+        public double Size{ get; set; }
 
         public Hook() { }
 
-        public Hook(int id, float size) {  Id = id; Size = size; }
+        public Hook(int id, double size) {  Id = id; Size = size; }
     }
 }
